moved hooks and utils to src and moved hook-call
