+++ frontend/altinn-support-dashboard.client/src/utils/utils.ts
